requirements.txt
sentence-transformers==2.2.2
llama-index==0.8.59
pypdf==3.16.4
faiss-cpu==1.7.4
gradio==3.50.2
